Another big commit:
1) Replace the "RadioBox" control with a ListCntrl control, for the call
list.  This has expandable columns, and displays each data type in it's
own column.

2) Some locking for the library: try to prevent us from doing the
processing loop while moving the calls around from the front-end.

3) REGISTRATION, and better incoming call support.

Incoming calls now are not auto-answered, but are answered when you
select them.

You can register with an IAX server (tested for IAX1 only so far).