The yankees are killing the bosox (or, they were when I started writing this), and the sox are making it interesting.
I figured I'd commit this stuff now, because if the yanks lose, I'll be
really pissed off.  Mussine had a perfect game through 6+.

Anyway, this is what it looks like: multiple codec support.

libiax2 doesn't yet provide a way for us to say we support multiple codecs,
but prefer one, but that's easy to change.  ulaw support added, but not yet
tested for correctness (I'm working on a remote machine, and I tried it, but
I can't hear the headset from 20 miles away).  gsm support still works.

The codec API is a little ugly, as the codec structure supports
bidirectional operation (encode and decode), but I use a separate one for
encoding and decoding right now.  I'll change it later so that it only does
that in the unlikely case that there's two codecs used.

speex support should be easy to add now (like an hour or so of work), and
iLBC as well.  I probably _wont_ include iLBC source in the archive here
initially, but I'll include the glue, and the Makefule-fu.