A bunch of new stuff:
First, the small stuff, and side-effects:

1) Added new API iaxc_set_networking(iaxc_sendto_t st, iaxc_recvfrom_t rf) ;
    application-defined networking; give substiture sendto and recvfrom
    functions, must be called before iaxc_initialize!

    This is useful if you want to handle the UDP I/O yourself, in your
    own application.  If you don't do anything, it will still be handled
    internally.

2) IAX1 support in iaxclient is now totally unusable;  We should remove
        the code now as well.
3) in audio/portaudio, there's a couple of commented fprintf statements
        for diagnosing output under/overflow 4) libiax2 now
        understands/parses incoming CNG frames.

Now, the big thing:

New jitterbuffer support.

This is the new jitterbuffer described at
http://www.voip-info.org/tiki-index.php?page=Asterisk%20new%20jitterbuffer

This is /NOT COMPLETE/ and it is disabled by default.  But (after
disabling some debugging), it is probably already usable.  It is being
committed now mainly for evaluation and enhancement by developers
interested in it.

To use it, define USE_NEWJB in lib/Makefile and recompile.

This jitterbuffer features:

1) PLC support:  Lost frames are interpolated, using either their own
        internal interpolation routines (speex, iLBC), or a
        quick-and-dirty interpolation routine (uLaw, aLaw, GSM).  The
        interpolation and support for these frames types is in iaxclient
        itself, but will only be used with the new jitterbuffer.

2) New jitterbuffer itself:  This is needed
        to use PLC.  It also uses interpolation to hide jitterbuffer growth.

The jitterbuffer itelf is in jitterbuf.[ch], and contains two different
implementations:  A literal histogram approach, and a statistical
approach by Jean-Marc Valin from libspeex.  Defining USE_SPEEX_JB uses
the libspeex methodology.