Remove select() loop in iaxc_service_network, and fold "do_iax_event" into iaxc_service_network.
The loop was unnecessary, because we loop again anyway, and the select breaks if "netfd" isn't
the right fd (as in, using iaxc_set_networking).  In that case, netfd=0, and if you have data on stdin,
we spinloop, because we never read stdin..

Generally, this just cleans up useless cruft.