,
Someone has reported a problem in my PLC code, now being used in iaxclient. He is using 2ms packetisation (I
don't know what he is doing with packets so small) and he can get crashes. His workaround is this:

[patch]

Now that is not a real solution, but at least it makes things reliable. The reason it is not a real solution
is it does not fill in proper pitch cycles when the overlap is bigger than "len". The audio quality will
suffer as a result.

I never considered such small packets when implementing this code. If there is a broad need for this to be
rock solid for any packet size I guess the code needs to be reworked a bit so the overlap can spread across
mulitple packets.

Regards,
Steve