From: Panfilov Dmitry <dima@bdpu.org> (with modifications by stevek)
Dmitry Panfilov has contributed a basic native ALSA driver for Linux.
Not all features are supported with this driver.  It is not compiled in
by default, because this would add alsa libraries to the build and link
dependencies -- and we don't have a good way of communicating that to
applications (like pkg-config stuff, etc). To use it, though, you just
need to change AUDIO_ALSA=0 to AUDIO_ALSA=1 in the Makefile.