okay, this is where we want to be.
first, i've reconciled changes to libiax2/jitterbuf.* with copies 
that once live in up in lib.

second, target_extra is now configurable in the jitterbuffer.  
it defaults to the #define'd JB_TARGET_EXTRA, but can be set 
via the jb_setconf() call since a target_extra has been added
to the jb_conf struct.