Fix race condition affected events containing audio data.
When IAXC_EVENT_AUDIO events were posted, the associated data buffer was a shared, static buffer. This made it so queued events would crap on previous events' data and if multiple audio events were posted, the data would be duplicated for all audio frames. This caused any application using these audio events to get bytes that were periodically bogus.

The solution is to treat the audio data the same way we treat video. That is, we allocate a chunk of memory and give each iaxc_event its own copy of the data.