Apply patch from stevebvt -- thank you!
This fixes an issue with portmixer and certain usb audio devices, for example the vitelix vx200 and polycom cx100. These devices show up not as speaker destinations but as undefined. This patch searches for volume controls preferring speaker destinations, then headphone destinations, and finally undefined destinations.

This patch also makes it so we find the wave output volume mixer control in Px_OpenMixer(). We then use the same mixerGetControlDetails()/mixerSetControlDetails() API to manipulate the volume for both Px_Get/SetPCMOutputVolume() and Px_Get/SetMasterVolume(). We used to use the waveOutGet/SetVolume() functions to manipulate the PCM volume.

Thanks again to stevebvt for this carefully constructed patch.